requirements.txt
# Analytics Agent v2.0 - Dependências Completas

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
numpy==1.24.3

# Presentation Generation
python-pptx==0.6.23

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# Logging and Configuration
loguru==0.7.2
python-dotenv==1.0.0

# AI and Machine Learning
openai==1.102.0
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.2

# Natural Language Processing
nltk==3.9.1
textblob==0.19.0
regex==2025.8.29

# Utilities
distro==1.9.0
jiter==0.10.0
threadpoolctl==3.6.0
